pyproject.toml
[project]
name = "audio-processor"
version = "0.1.0"
description = "Audio extraction and processing SPA with tunnel/reverb effects"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "loguru>=0.7.0",
    "yt-dlp>=2024.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
audio-processor = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
